Added Sryheni Pathing

+++ Implementations/SimpleConnection.cs
@@ -7,11 +7,11 @@ namespace FedoraDev.NodeEditor.Implementations
 	{
 		public INode NodeA { get => _nodeA; set => _nodeA = value; }
 		public INode NodeB { get => _nodeB; set => _nodeB = value; }
-		public float Distance { get => _distance; set => _distance = value; }
+		public int Distance { get => _distance; set => _distance = value; }
 
 		[SerializeField, ReadOnly] INode _nodeA;
 		[SerializeField, ReadOnly] INode _nodeB;
-		[SerializeField, ReadOnly] float _distance = 1;
+		[SerializeField, ReadOnly] int _distance = 1;
 
 		public INode GetOtherNode(INode node) => NodeA == node ? NodeB : NodeA;
 		public IConnection Produce(IFactory _factory) => new SimpleConnection();
